MM_30482 :  Refactored Handlers to use Typescript Mixins
